requirements.txt
azure-keyvault-secrets==4.3.0
azure-keyvault-keys==4.4.0
azure-identity==1.7.0
